requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.13.1
fonttools==4.58.3
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.3
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
openlocationcode==1.0.1
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.0.0
pip==25.1.1
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.154
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1